examples/pg_vector_example/requirements.txt
pgvector
psycopg2
python-dotenv